Crate blog

This is the rust doc for the blog library - what the blog binary depends on to get posts, and convert markdown to html properly.

For documentation on using the blog binary, see the main readme. The entry point is main.rs.

Strategy

All markdown files are read at startup by globbing for README.md in the posts/ directory. These files are converted to HTML and stored in a map. It's also linked up with a data.json and all links are converted from relative links to something with a /static prefix.

Summaries are created by somewhat naively trying to find the first paragraph that is not a raw heading or an image.

Serving

The main blog engine would be in the binary, and that's the only thing that is depending on rocket and its ecosystem.

Structs

DataBase
All the data we have in posts/

MetaData
The metadata representation of the data.json files

Post
The full internal representation of a post subfolder

Functions

load_posts
A one time sequential loader of all posts from the posts folder